+++ packages/azure-signing-manager/jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  displayName: 'azure-signing-manager',
+  preset: '../../jest.preset.js',
+  transform: {
+    '^.+\\.(ts)$': [
+      'ts-jest',
+      {
+        tsconfig: '<rootDir>/tsconfig.spec.json',
+      },
+    ],
+  },
+  testEnvironment: 'node',
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../coverage/packages/azure-signing-manager',
+};