+++ packages/azure-signing-manager/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/azure-signing-manager';
+export * from './types';